requirements.txt
# VisionResearch Project Requirements
# Action-Video Mismatch Experiment

# Core dependencies
numpy>=1.21.0
pillow>=8.0.0

# Data processing
pandas>=1.3.0
scipy>=1.7.0

# Deep learning (for VGGT oracle)
torch>=1.9.0
torchvision>=0.10.0

# Visualization
matplotlib>=3.4.0

# Computer vision fallback (OpenCV pose estimator)
opencv-python>=4.5.0

# Optional: for better plots
seaborn>=0.11.0
